requirements.txt
aiosmtplib==2.0.2
alabaster==0.7.13
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
Babel==2.12.1
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cloudinary==1.33.0
coverage==7.2.7
cryptography==41.0.2
dnspython==2.4.1
docutils==0.20.1
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.100.0
fastapi-limiter==0.1.5
fastapi-mail==1.4.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
imagesize==1.4.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
orjson==3.9.2
packaging==23.1
passlib==1.7.4
Pillow==10.0.0
pluggy==1.2.0
psycopg2-binary
pyasn1==0.5.0
pycparser==2.21
pydantic==2.0.3
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
pydantic_core==2.3.0
Pygments==2.15.1
pypng==0.20220715.0
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
qrcode==7.4.2
redis==4.6.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
Sphinx==7.1.1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqladmin==0.13.0
SQLAlchemy==2.0.19
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
ujson==5.8.0
urllib3==1.26.16
uvicorn==0.23.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
WTForms==3.0.1
